requirements.txt
pandas==2.2.0
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.1
feedparser==6.0.10
beautifulsoup4==4.12.2
schedule==1.2.0
streamlit==1.28.1
plotly==5.17.0
openai==1.59.6
